Add `DeleteRecentPlaylists` example
This is useful for cleaning up the results of the other example.

The original plan was to delete all playlists created in the last `n` seconds, but playlists don't actually have any sort of `added_at` field (tracks do).

We assume that `getMyPlaylists` returns the most recent first. Which seems to be the case. But it isn't disastrous if it isn't, seeing as we check for confirmation anyway.