Consistently define types in types package
This makes the core types package more useful to non-servant clients. Plus, the servant package can have a simple consistent structure.

The reason this wasn't already the case is that Spotify used to explictily document the types used in the API, so our types package originally mapped precisely to that. But now they just repeat the structures in each endpoint definition...